+++ src/common/ui/display_bagl.c
@@ -824,10 +824,75 @@ UX_FLOW(ux_sign_transfer_to_public,
         &ux_sign_flow_shared_sign,
         &ux_sign_flow_shared_decline);
 
-
 void uiSignTransferToPublicDisplay(volatile unsigned int *flags) {
     ux_flow_init(0, ux_sign_transfer_to_public, NULL);
     *flags |= IO_ASYNCH_REPLY;
+}
+
+// Sign Transfer with Schedule
+const ux_flow_step_t *ux_sign_scheduled_amount_transfer[8];
+static signTransferWithScheduleContext_t *ctx_sign_transfer_with_schedule =
+    &global.withDataBlob.signTransferWithScheduleContext;
+
+// UI definitions for displaying the transaction contents of the first packet for verification
+// before continuing to process the scheduled amount pairs that will be received in separate
+// packets.
+UX_STEP_NOCB(ux_scheduled_transfer_initial_flow_1_step,
+             bnnn_paging,
+             {.title = "Recipient",
+              .text = (char *) global.withDataBlob.signTransferWithScheduleContext.displayStr});
+UX_STEP_VALID(ux_scheduled_transfer_initial_flow_2_step,
+              nn,
+              sendSuccessNoIdle(),
+              {"Continue", "with transaction"});
+
+// UI definitions for displaying a timestamp and an amount of a scheduled transfer.
+UX_STEP_NOCB(ux_sign_scheduled_transfer_pair_flow_0_step,
+             bnnn_paging,
+             {"Release time (UTC)",
+              (char *) global.withDataBlob.signTransferWithScheduleContext.displayTimestamp});
+UX_STEP_NOCB(ux_sign_scheduled_transfer_pair_flow_1_step,
+             bnnn_paging,
+             {"Amount",
+              (char *) global.withDataBlob.signTransferWithScheduleContext.displayAmount});
+UX_STEP_CB(ux_sign_scheduled_transfer_pair_flow_2_step,
+           nn,
+           processNextScheduledAmount(ctx_sign_transfer_with_schedule->buffer),
+           {"Show", "next release"});
+UX_FLOW(ux_sign_scheduled_transfer_pair_flow,
+        &ux_sign_scheduled_transfer_pair_flow_0_step,
+        &ux_sign_scheduled_transfer_pair_flow_1_step,
+        &ux_sign_scheduled_transfer_pair_flow_2_step);
+
+UX_FLOW(ux_sign_scheduled_transfer_pair_flow_sign,
+        &ux_sign_scheduled_transfer_pair_flow_0_step,
+        &ux_sign_scheduled_transfer_pair_flow_1_step,
+        &ux_sign_flow_shared_sign,
+        &ux_sign_flow_shared_decline);
+
+void startInitialScheduledTransferDisplay(bool displayMemo) {
+    uint8_t index = 0;
+
+    ux_sign_scheduled_amount_transfer[index++] = &ux_sign_flow_shared_review;
+    ux_sign_scheduled_amount_transfer[index++] = &ux_sign_flow_account_sender_view;
+    ux_sign_scheduled_amount_transfer[index++] = &ux_scheduled_transfer_initial_flow_1_step;
+
+    if (displayMemo) {
+        ux_sign_scheduled_amount_transfer[index++] = &ux_display_memo_step_nocb;
+    }
+
+    ux_sign_scheduled_amount_transfer[index++] = &ux_scheduled_transfer_initial_flow_2_step;
+
+    ux_sign_scheduled_amount_transfer[index++] = FLOW_END_STEP;
+    ux_flow_init(0, ux_sign_scheduled_amount_transfer, NULL);
+}
+
+void uiSignScheduledTransferPairFlowSignDisplay(void) {
+    ux_flow_init(0, ux_sign_scheduled_transfer_pair_flow_sign, NULL);
+}
+
+void uiSignScheduledTransferPairFlowDisplay(void) {
+    ux_flow_init(0, ux_sign_scheduled_transfer_pair_flow, NULL);
 }
 
 #endif
